Multi-Column Keys
The B-Tree implementation now includes keys that span multiple columns. When
creating a tree, an array of integers indicating which columns to use, and in
which order, is expected. The tree's Get method expects the key's width to
match the tree's key width. The GetRange, however, can take a key of any width.